ml-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
pymongo==4.6.1
python-dotenv==1.0.0
joblib==1.3.2
pydantic==2.5.3
pyarrow==15.0.0
xgboost==2.0.3
lightgbm==4.3.0
shap==0.44.1
statsmodels==0.14.1
python-multipart==0.0.9
mapie==0.8.2
gymnasium==0.29.1
stable-baselines3==2.2.1
mlflow==2.10.0
redis==5.0.1
apscheduler==3.10.4
requests==2.31.0
prometheus-fastapi-instrumentator==6.1.0
pytrends==4.9.2
structlog==24.1.0
scipy==1.12.0
# MLflow experiment tracking